# go/GinApiGateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "包含所有 api 的用法.",
        "title": "Prizoo Api Document",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/activity": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "獲取所有活動的資料",
                "responses": {
                    "200": {
                        "description": "{\"data\": \"活動內容\"}",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "新建一個活動",
                "parameters": [
                    {
                        "type": "string",
                        "description": "活動名稱",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "活動細節介紹",
                        "name": "detial",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "活動開始時間(unix time stamp)",
                        "name": "startTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "活動結束時間(unix time stamp)",
                        "name": "endTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "提供活動的廠商名稱",
                        "name": "partner",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "獎品名稱",
                        "name": "prizeName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "獎品的圖片",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "加碼的獎品名稱",
                        "name": "addPrizeName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "加碼獎品的圖片",
                        "name": "addPrizeImage",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "多少票券加入之後開啟加碼",
                        "name": "addPrizeTicketAmount",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/activity.CreateActivityResponse"
                        }
                    },
                    "400": {
                        "description": "輸入錯誤",
                        "schema": {
                            "$ref": "#/definitions/activity.CreateActivityResponse"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/activity.CreateActivityResponse"
                        }
                    }
                }
            }
        },
        "/activity/history/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "查看特定活動 id 的修改歷史記錄",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要查看的活動 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\": \"歷史記錄陣列\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/activity/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "獲取特定 id 的活動",
                "parameters": [
                    {
                        "type": "string",
                        "description": "活動 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\": \"活動內容\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"message\": \"錯誤訊息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"message\": \"獲取失敗，請重試\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "刪除特定 id 的活動",
                "parameters": [
                    {
                        "type": "string",
                        "description": "活動 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/activity.DeleteActivityResponse"
                        }
                    },
                    "400": {
                        "description": "錯誤的 id",
                        "schema": {
                            "$ref": "#/definitions/activity.DeleteActivityResponse"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/activity.DeleteActivityResponse"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "更新指定 id 的活動內容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要更新的活動 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "活動名稱",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "活動細節介紹",
                        "name": "detial",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "活動開始時間(unix time stamp)",
                        "name": "startTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "活動結束時間(unix time stamp)",
                        "name": "endTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "提供活動的廠商名稱",
                        "name": "partner",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "獎品名稱",
                        "name": "prizeName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "獎品的圖片",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "加碼的獎品名稱",
                        "name": "addPrizeName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "多少票券加入之後開啟加碼",
                        "name": "addPrizeTicketAmount",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/activity.UpdateActivityResponse"
                        }
                    },
                    "400": {
                        "description": "錯誤的 id",
                        "schema": {
                            "$ref": "#/definitions/activity.UpdateActivityResponse"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/activity.UpdateActivityResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "activity.CreateActivityResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "activity.DeleteActivityResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "activity.UpdateActivityResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}